+++ packages/components-library/src/main.ts
@@ -1,7 +1,9 @@
-import CartSelectionToast from './components/CartSelectionToast.vue'
+import CartSelectionToast from './components/ui/CartSelectionToast.vue'
+import CheckboxFilter from './components/filters/CheckboxFilter.vue'
+import StringFilter from './components/filters/StringFilter.vue'
 
-export { CartSelectionToast }
+export { CartSelectionToast, CheckboxFilter, StringFilter }
 
 export default {
-  CartSelectionToast
+  CartSelectionToast, CheckboxFilter, StringFilter
 }