+++ packages/data-explorer/tests/unit/views/CardView.spec.ts
@@ -13,13 +13,12 @@ describe('CardView.vue', () => {
 
   beforeEach(() => {
     state = {
-      shoppingFilter: false,
       tableMeta: { idAttribute: { name: 'tableID' }, labelAttribute: { name: 'label' }, attributes: [{ type: 'compound' }, { type: 'string' }, { type: 'string' }] },
       dataDisplayLayout: 'ClipboardView',
-      shoppedEntityItems: ['1', '3']
+      selectedItemIds: ['1', '3']
     }
     mutations = {
-      toggleShoppingItems: jest.fn()
+      toggleSelectedItems: jest.fn()
     }
     actions = {
       fetchRowDataLabels: jest.fn(),
@@ -80,10 +79,9 @@ describe('CardView.vue', () => {
 
   it('returns id if label attribute is not defined', () => {
     state = {
-      shoppingFilter: false,
       tableMeta: { idAttribute: { name: 'tableID' } },
       dataDisplayLayout: 'ClipboardView',
-      shoppedEntityItems: [1, 3]
+      selectedItemIds: [1, 3]
     }
     store = new Vuex.Store({
       state, mutations, getters, actions