+++ packages/data-explorer/tests/unit/views/DataView.spec.ts
@@ -16,6 +16,10 @@ describe('DataView.vue', () => {
       settingsTable: 'settings-table',
       dataDisplayLayout: 'cards',
       tableSettings: {},
+      showSelected: false,
+      selectedItemIds: [],
+      tableMeta: {},
+      tableData: {},
       filters: {
         hideSidebar: false,
         definition: [],
@@ -29,6 +33,9 @@ describe('DataView.vue', () => {
     }
     mutations = {
       setSearchText: jest.fn(),
+      setHideFilters: jest.fn(),
+      setShowSelected: jest.fn(),
+      setSelectedItems: jest.fn(),
       setFilterSelection: jest.fn()
     }
     actions = {
@@ -50,6 +57,7 @@ describe('DataView.vue', () => {
       wrapper = shallowMount(DataView, { store, localVue })
       wrapper.setData({ searchText: 'test' })
     })
+
     it('should mutate the value to the store', () => {
       expect(mutations.setSearchText).toHaveBeenCalled()
     })
@@ -60,15 +68,60 @@ describe('DataView.vue', () => {
     beforeEach(() => {
       wrapper = shallowMount(DataView, { store, localVue })
     })
+
     it('should clear the search text if search is not part of the filter', () => {
       const newSelections = {}
       wrapper.vm.saveFilterState(newSelections)
       expect(mutations.setSearchText).toHaveBeenCalled()
     })
+
     it('should not clear the search text if search is part of the filter', () => {
       const newSelections = { _search: 'mock selection' }
+      // @ts-ignore
       wrapper.vm.saveFilterState(newSelections)
       expect(mutations.setSearchText).not.toHaveBeenCalled()
+    })
+  })
+
+  describe('cartSelectionToast component', () => {
+    it('should hide filters and change to selection list when openSelectionList is called', () => {
+      const wrapper = shallowMount(DataView, { store, localVue })
+      // @ts-ignore
+      wrapper.vm.openSelectionList()
+      expect(mutations.setShowSelected).toHaveBeenCalledWith(state, true)
+      expect(mutations.setHideFilters).toHaveBeenCalledWith(state, true)
+    })
+
+    it('should find a name to use as display label', () => {
+      const wrapper = shallowMount(DataView, { store, localVue })
+      // @ts-ignore
+      expect(wrapper.vm.displayName).toEqual('name')
+
+      store.state.tableMeta = { labelAttribute: { name: 'label' } }
+      const wrapper2 = shallowMount(DataView, { store, localVue })
+      // @ts-ignore
+      expect(wrapper2.vm.displayName).toEqual('label')
+    })
+
+    it('should add names to display to the items in the shoppingcart', () => {
+      const wrapper = shallowMount(DataView, { store, localVue })
+      // @ts-ignore
+      expect(wrapper.vm.handleSelectionItems).toEqual([])
+
+      store.state.tableMeta = { labelAttribute: { name: 'label' } }
+      store.state.tableData = { items: [{ label: 'item1', id: '1' }, { label: 'item2', id: '2' }] }
+      store.state.selectedItemIds = ['1', '2']
+
+      const wrapper2 = shallowMount(DataView, { store, localVue })
+      // @ts-ignore
+      expect(wrapper2.vm.handleSelectionItems).toEqual([{ name: 'item1', id: '1' }, { name: 'item2', id: '2' }])
+    })
+
+    it('should remove names to display from items before returning them', () => {
+      const wrapper = shallowMount(DataView, { store, localVue })
+      // @ts-ignore
+      wrapper.vm.handleSelectionItems = [{ name: 'item1', id: '1' }, { name: 'item2', id: '2' }]
+      expect(mutations.setSelectedItems).toHaveBeenCalledWith(state, ['1', '2'])
     })
   })
 
@@ -80,10 +133,12 @@ describe('DataView.vue', () => {
     })
 
     it('should add search to the active filter selection', () => {
+      // @ts-ignore
       expect(wrapper.vm.activeFilterSelections).toEqual({ _search: 'my search' })
     })
 
     it('should add search to the active filter selection', () => {
+      // @ts-ignore
       expect(wrapper.vm.filterDefinitions).toEqual([{
         type: 'string',
         label: 'search',